Move arbitrary impl to module
